Fixed several compiler warnings
- Warnings detected thanks to Xcode's compiler settings (more strict by
  default) and clang, warnings mostly but not only related to data sizes
  on 64 bit systems, that were silenced until now by very lax compiler
  settings.
- This also decreases by a great deal the amount of warnings produced by
  MSVC in x64 mode (for the adventurous ones who tried that)
- Also fixed (or silenced in case of false positives) the potential
  issues pointed out by the (awesome) clang static analyzer.
- Patch co-produced with Ind, I'm merging and committing in his place!

+++ src/common/malloc.c
@@ -669,7 +669,7 @@ void memmgr_report (int extra) {
 	struct {
 		const char *file;
 		unsigned short line;
-		unsigned int size;
+		size_t size;
 		unsigned int count;
 	} data[100];
 	memset(&data, 0, sizeof(data));
